move connection handling from server to library
