+++ library/connection.c
@@ -1,21 +1,21 @@
 /*
  * I/O routines for test server.
- * 
+ *
  * The idea is to avoid interfering with networks test. This enables to test
  * even with all network interfaces are shut down.
- * 
- * 
+ *
+ *
  * Copyright (C) 2014-2015 SUSE
- * 
+ *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, version 2.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License along
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
@@ -41,7 +41,7 @@
 #include <string.h>
 #include <stdint.h>
 
-#include "server.h"
+#include "connection.h"
 
 
 /*
@@ -396,4 +396,3 @@ twopence_conn_pool_poll(twopence_conn_pool_t *pool)
 
 	return !!pool->connections;
 }
-
