+++ library/connection.h
